+++ goldens/circular-deps/packages.json
@@ -264,21 +264,6 @@
     "packages/core/src/render3/view_ref.ts",
     "packages/core/src/linker/view_container_ref.ts",
     "packages/core/src/render3/di.ts",
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/injector.ts",
-    "packages/core/src/di/r3_injector.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/metadata/ng_module.ts"
-  ],
-  [
-    "packages/core/src/application_ref.ts",
-    "packages/core/src/application_tokens.ts",
-    "packages/core/src/linker/component_factory.ts",
-    "packages/core/src/change_detection/change_detection.ts",
-    "packages/core/src/change_detection/change_detector_ref.ts",
-    "packages/core/src/render3/view_ref.ts",
-    "packages/core/src/linker/view_container_ref.ts",
-    "packages/core/src/render3/di.ts",
     "packages/core/src/di/injector.ts",
     "packages/core/src/di/r3_injector.ts",
     "packages/core/src/render3/definition.ts",
@@ -645,16 +630,6 @@
     "packages/core/src/application_ref.ts",
     "packages/core/src/linker/compiler.ts",
     "packages/core/src/render3/ng_module_ref.ts",
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/injector.ts",
-    "packages/core/src/di/r3_injector.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/metadata/ng_module.ts"
-  ],
-  [
-    "packages/core/src/application_ref.ts",
-    "packages/core/src/linker/compiler.ts",
-    "packages/core/src/render3/ng_module_ref.ts",
     "packages/core/src/di/injector.ts",
     "packages/core/src/di/r3_injector.ts",
     "packages/core/src/render3/definition.ts",
@@ -794,20 +769,6 @@
     "packages/core/src/render3/jit/environment.ts",
     "packages/core/src/render3/index.ts",
     "packages/core/src/render3/pipe.ts",
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/injector.ts",
-    "packages/core/src/di/r3_injector.ts",
-    "packages/core/src/render3/definition.ts",
-    "packages/core/src/metadata/ng_module.ts"
-  ],
-  [
-    "packages/core/src/application_ref.ts",
-    "packages/core/src/metadata/resource_loading.ts",
-    "packages/core/src/metadata/directives.ts",
-    "packages/core/src/render3/jit/directive.ts",
-    "packages/core/src/render3/jit/environment.ts",
-    "packages/core/src/render3/index.ts",
-    "packages/core/src/render3/pipe.ts",
     "packages/core/src/render3/definition.ts",
     "packages/core/src/metadata/ng_module.ts"
   ],
@@ -901,12 +862,27 @@
     "packages/core/src/view/services.ts"
   ],
   [
+    "packages/core/src/di.ts",
+    "packages/core/src/di/index.ts",
+    "packages/core/src/di/injectable.ts",
+    "packages/core/src/di/jit/injectable.ts",
+    "packages/core/src/di/jit/environment.ts",
+    "packages/core/src/di/injector_compatibility.ts",
+    "packages/core/src/di/injector.ts",
+    "packages/core/src/di/null_injector.ts"
+  ],
+  [
     "packages/core/src/di/injectable.ts",
     "packages/core/src/di/jit/injectable.ts"
   ],
   [
     "packages/core/src/di/injector_compatibility.ts",
     "packages/core/src/di/injector.ts"
+  ],
+  [
+    "packages/core/src/di/injector_compatibility.ts",
+    "packages/core/src/di/injector.ts",
+    "packages/core/src/di/null_injector.ts"
   ],
   [
     "packages/core/src/di/injector_compatibility.ts",
@@ -920,6 +896,10 @@
     "packages/core/src/render3/definition.ts",
     "packages/core/src/metadata/ng_module.ts",
     "packages/core/src/di/util.ts"
+  ],
+  [
+    "packages/core/src/di/injector_token.ts",
+    "packages/core/src/di/injector.ts"
   ],
   [
     "packages/core/src/di/injector.ts",
