+++ goldens/size-tracking/aio-payloads.json
@@ -12,7 +12,7 @@
     "master": {
       "uncompressed": {
         "runtime-es2015": 3037,
-        "main-es2015": 449483,
+        "main-es2015": 448796,
         "polyfills-es2015": 52415
       }
     }
