+++ packages/core/src/render3/pipe.ts
@@ -8,8 +8,7 @@
 
 import {WrappedValue} from '../change_detection/change_detection_util';
 import {PipeTransform} from '../change_detection/pipe_transform';
-import {setInjectImplementation} from '../di/injector_compatibility';
-
+import {setInjectImplementation} from '../di/inject_switch';
 import {getFactoryDef} from './definition';
 import {setIncludeViewProviders} from './di';
 import {RuntimeError, RuntimeErrorCode} from './error_code';
