+++ packages/core/test/bundling/forms/bundle.golden_symbols.json
@@ -834,6 +834,9 @@
     "name": "createLView"
   },
   {
+    "name": "createNodeInjector"
+  },
+  {
     "name": "createPlatformFactory"
   },
   {
