+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -1074,6 +1074,9 @@
     "name": "createNewSegmentGroup"
   },
   {
+    "name": "createNodeInjector"
+  },
+  {
     "name": "createPlatformFactory"
   },
   {
