+++ packages/core/test/bundling/todo/bundle.golden_symbols.json
@@ -273,6 +273,9 @@
     "name": "createLView"
   },
   {
+    "name": "createNodeInjector"
+  },
+  {
     "name": "createTView"
   },
   {
